refactor: extract core library and add spectre cli
