+++ global.json
@@ -1,6 +1,6 @@
 {
   "sdk": {
-    "version": "9.0.0",
+    "version": "8.0.119",
     "rollForward": "latestMinor",
     "allowPrerelease": false
   }
