--- src/ParameterMapping.cs
@@ -1,8 +1,0 @@
-﻿namespace LinqToSqlMetadataExtractor;
-
-class ParameterMapping
-{
-    public string Name { get; set; }
-    public string Type { get; set; }
-    public bool IsNullable { get; set; }
-}
