--- src/TableMapping.cs
@@ -1,7 +1,0 @@
-﻿namespace LinqToSqlMetadataExtractor;
-
-class TableMapping
-{
-    public string Name { get; set; }
-    public string EntityType { get; set; }
-}
